requirements.txt
discord.py
aiohttp
logging
